requirements.txt
# general
jsonlines
pandas
python-dotenv

# dataset
requests
datasets
gdown

# lm inference
vllm
openai

# longwiki
segtok
sentence-transformers==3.3.1

# refusal
ratelimit
retry